restructure file hirarchy
