+++ src/Firebase/FirebaseFunctions/DBfuncs.tsx
@@ -1,7 +1,7 @@
 import firebase from "firebase/app";
 import "firebase/database";
 import { dataref } from "../FirebaseConfig/firebase";
-import { Job, Candidate, CandidateJobStatus, Recruiter } from "./types";
+import { getFilteredJobs } from "./Job";
 
 /**
  * Prints the data located at the given path in the Firebase Realtime Database.
@@ -50,209 +50,6 @@ async function appendToDatabase(obj: any, path: string) {
 		await database.ref(path).push(obj);
 	} catch (error) {
 		console.error("Error adding object to database:", error);
-	}
-}
-/* Jobs functions */
-/**
- * Retrieves a list of all jobs from the Firebase Realtime Database.
- * @async
- * @function getJobsFromDatabase
- * @returns {Promise<Job[]>} A promise that resolves to an array of Job objects.
- * @throws {Error} If there is an error fetching the jobs from the database.
- */
-async function getJobsFromDatabase(): Promise<Job[]> {
-	const database = dataref;
-	try {
-		const snapshot = await database.ref("/Jobs").once("value");
-		const jobsData = snapshot.val();
-		const jobs: Job[] = [];
-		for (const i in jobsData) {
-			const job = jobsData[i];
-			jobs.push(job);
-		}
-		return jobs;
-	} catch (error) {
-		console.error(error);
-		throw new Error("Failed to fetch jobs from database.");
-	}
-}
-/**
- * Retrieves a list of all jobs from the database.
- * @returns {Promise<Array>} A promise that resolves to an array of job objects.
- */
-async function getJobs() {
-	return getJobsFromDatabase();
-}
-
-/**
- * Filters the list of jobs based on the given attributes and values, and sorts the result
- * @param {string[]} [attributes=[]] - An array of attributes name to filter by.
- * @param {string[]} [values=[]] - An array of values to filter by,
- *  make sure the order of the values is correspond to the order of attributes array.
- * 	sendd all values as string for boolean value use 'true' or 'false' number as '123' and scope '50,100'
- * @param {string} [sortBy=""] - The attribute to sort the jobs by.
- * @returns {Promise<Job[]>} - A promise that resolves to an array of Job objects that match the filter criteria and sorted by desired attribute.
- */
-export async function getFilteredJobs(attributes: string[] = [], values: string[] = [], sortBy: string = "") {
-	if (attributes.length !== values.length) {
-		console.log("the attributes length not match to values length")
-		return [];
-	}
-	let jobs = await getJobs();
-	//filtering
-	let i = attributes.indexOf("jobNumber");
-	if (i >= 0) {
-		jobs = jobs.filter(job => job._jobNumber.toString() === values[i])
-	}
-	i = attributes.indexOf("title");
-	if (i >= 0) {
-		jobs = jobs.filter(job => job._title === values[i])
-	}
-	i = attributes.indexOf("role");
-	if (i >= 0) {
-		jobs = jobs.filter(job => job._role === values[i])
-	}
-	i = attributes.indexOf("scope");
-	if (i >= 0) {
-		let scope = values[i].split(",").map(num => parseInt(num))
-		jobs = jobs.filter((job) => job._scope[0] <= scope[0] || scope[1] <= job._scope[1]);
-	}
-	i = attributes.indexOf("description");
-	if (i >= 0) {
-		jobs = jobs.filter(job => job._description.includes(values[i]));
-	}
-	i = attributes.indexOf("open");
-	if (i >= 0) {
-		jobs = jobs.filter(job => job._open.toString() === values[i])
-	}
-	i = attributes.indexOf("highPriority");
-	if (i >= 0) {
-		jobs = jobs.filter(job => job._highPriority.toString() === values[i])
-	}
-	i = attributes.indexOf("sector");
-	if (i >= 0) {
-		jobs = jobs.filter(job => job._sector === values[i])
-	}
-	//sorting
-	if (sortBy === "title")
-		return jobs.sort(compareByTitle);
-	if (sortBy === "jobNumber")
-		return jobs.sort(compareByJobNumber);
-	if (sortBy === "role")
-		return jobs.sort(compareByRole);
-	if (sortBy === "scope")
-		return jobs.sort(compareByScope);
-	if (sortBy === "region")
-		return jobs.sort(compareByRegion);
-	if (sortBy === "sector")
-		return jobs.sort(compareBySector);
-	if (sortBy === "open")
-		return jobs.sort(compareByOpen);
-	if (sortBy === "hightPriority")
-		return jobs.sort(compareByHighPriority);
-	if (sortBy === "creationDate")
-		return jobs.sort(compareByCreationDate);
-	if (sortBy === "views")
-		return jobs.sort(compareByViews);
-	return jobs;
-}
-/* compare function for sort */
-function compareByTitle(a: Job, b: Job): number {
-	return a._title.localeCompare(b._title);
-}
-
-function compareByJobNumber(a: Job, b: Job): number {
-	return a._jobNumber - b._jobNumber;
-}
-
-function compareByRole(a: Job, b: Job): number {
-	return a._role.localeCompare(b._role);
-}
-
-function compareByScope(a: Job, b: Job): number {
-	return a._scope[1] - b._scope[1];
-}
-
-function compareByRegion(a: Job, b: Job): number {
-	return a._region.localeCompare(b._region);
-}
-
-function compareBySector(a: Job, b: Job): number {
-	return a._sector.localeCompare(b._sector);
-}
-
-function compareByOpen(a: Job, b: Job): number {
-	if (b._open === a._open)
-		return 0;
-	if (b._open === true)
-		return 1;
-	return -1;
-}
-
-function compareByHighPriority(a: Job, b: Job): number {
-	if (b._highPriority === a._highPriority)
-		return 0;
-	if (b._highPriority === true)
-		return 1;
-	return -1;
-}
-
-function compareByViews(a: Job, b: Job): number {
-	return b._views - a._views;
-}
-function compareByCreationDate(a: Job, b: Job): number {
-	if (b._creationDate > a._creationDate)
-		return 1;
-	if (b._creationDate < a._creationDate)
-		return -1;
-	return 0;
-}
-async function getCandidatesFromDatabase(): Promise<Candidate[]> {
-	const database = dataref;
-	try {
-		const snapshot = await database.ref("/Candidates").once("value");
-		const candidatesData = snapshot.val();
-		const candidates: Candidate[] = [];
-		for (const candidateId in candidatesData) {
-			const candidate = candidatesData[candidateId];
-			candidates.push(candidate);
-		}
-		return candidates;
-	} catch (error) {
-		console.error(error);
-		throw new Error("Failed to fetch candidates from database.");
-	}
-}
-async function getCandidateJobStatusFromDatabase(): Promise<CandidateJobStatus[]> {
-	const database = dataref;
-	try {
-		const snapshot = await database.ref("/CandidatesJobStatus").once("value");
-		const candidateJobStatusData = snapshot.val();
-		const candidateJobStatuses: CandidateJobStatus[] = [];
-		for (const candidateJobStatusId in candidateJobStatusData) {
-			const candidateJobStatus = candidateJobStatusData[candidateJobStatusId];
-			candidateJobStatuses.push(candidateJobStatus);
-		}
-		return candidateJobStatuses;
-	} catch (error) {
-		console.error(error);
-		throw new Error("Failed to fetch candidate job statuses from database.");
-	}
-}
-async function getRecruitersFromDatabase(): Promise<Recruiter[]> {
-	const database = dataref;
-	try {
-		const snapshot = await database.ref("/Recruiters").once("value");
-		const recruitersData = snapshot.val();
-		const recruiters: Recruiter[] = [];
-		for (const recruiterId in recruitersData) {
-			const recruiter = recruitersData[recruiterId];
-			recruiters.push(recruiter);
-		}
-		return recruiters;
-	} catch (error) {
-		console.error(error);
-		throw new Error("Failed to fetch recruiters from database.");
 	}
 }
